Serializing the test cases for connection resource (#10491)
[upstream:48d39391e2fc2b2916ee99087c432ba3bdb5921a]
